wNAF 5 precomputation for p384
This is used for both the base point, and any other point
multiplication, equally.  It improves the performance by about
80-90%.  We comfortably beat rustcrypto and ring, but are short
compared to aws-lc-rs.

We _could_ duplicate the wNAF 7 precomputation for the base point
from p256, but that would take 222KB of tables which seems like
a poor trade-off given the popularity of p384 right now.