Rework EC private->public key accessor
The `PublicKey` types are intended for peer public keys
(where you're about to do a multiplication with them,
for ECDH or ECDSA verification).  Therefore, they precompute
some multiples to prepare for this.

`PrivateKey::public_key()` was doing this too, but this is
_our_ public key, so the precomputation was a waste.

Avoid that by providing uncompressed encoding (and, for ECDSA
signing, x_scalar) accessors.