Implement Error for graviola::Error
This makes it possible to use the `?` operator with `graviola::Error` in
application code that uses e.g. `anyhow`.

The required implementations of `core::fmt::Display` are simply based on
the docs for each variant, converted to the standard lowercase,
non-punctuated style.

This currently used `std::error::Error` instead of `core::error::Error`
because the latter would mean a `rust-version` bump up to 1.81.
Depending on what happens with #57, this could perhaps use some feature
flag(s) to conditionally use `std` or `core`. Or hide the entire trait
impl behind a feature flag.